// src/codegen/src/bitway/tss/tx.ts
export interface MsgCompleteDKG {
  /** sender */
  sender: string;
  /** DKG request id */
  id: bigint;
  /** public keys generated by DKG */
  pubKeys: string[];
  /** participant consensus pub key */
  consensusPubkey: string;
  /** hex encoded participant signature */
  signature: string;
}
export interface MsgCompleteDKGProtoMsg {
  typeUrl: "/bitway.tss.MsgCompleteDKG";
  value: Uint8Array;
}
/** MsgCompleteDKG is the Msg/CompleteDKG request type. */
export interface MsgCompleteDKGAmino {
  /** sender */
  sender?: string;
  /** DKG request id */
  id?: string;
  /** public keys generated by DKG */
  pub_keys?: string[];
  /** participant consensus pub key */
  consensus_pubkey?: string;
  /** hex encoded participant signature */
  signature?: string;
}
// ...
export interface MsgSubmitSignatures {
  sender: string;
  id: bigint;
  signatures: string[];
}
// ...
export interface MsgRefresh {
  /** authority is the address that controls the module (defaults to x/gov unless overwritten). */
  authority: string;
  /** list of DKGs corresponding to key shares to be refreshed */
  dkgIds: bigint[];
  /** removed participant set */
  removedParticipants: string[];
  /** new threshold set corresponding to the DKGs */
  thresholds: number[];
  /** timeout duration per DKG refreshing */
  timeoutDuration: Duration;
}
// ...
export interface MsgCompleteRefreshing {
  /** sender */
  sender: string;
  /** request id */
  id: bigint;
  /** participant consensus pub key */
  consensusPubkey: string;
  /** hex encoded participant signature */
  signature: string;
}
// ...
export interface MsgUpdateParams {
  /** authority is the address that controls the module (defaults to x/gov unless overwritten). */
  authority: string;
  /**
   * params defines the x/tss parameters to be updated.
   * 
   * NOTE: All parameters must be supplied.
   */
  params: Params;
}
// ...
export const MsgCompleteDKG = {
  typeUrl: "/bitway.tss.MsgCompleteDKG",
  encode(message: MsgCompleteDKG, writer: BinaryWriter = BinaryWriter.create()): BinaryWriter {
    if (message.sender !== "") {
      writer.uint32(10).string(message.sender);
    }
    if (message.id !== BigInt(0)) {
      writer.uint32(16).uint64(message.id);
    }
    for (const v of message.pubKeys) {
      writer.uint32(26).string(v!);
    }
    if (message.consensusPubkey !== "") {
      writer.uint32(34).string(message.consensusPubkey);
    }
    if (message.signature !== "") {
      writer.uint32(42).string(message.signature);
    }
    return writer;
  },
  decode(input: BinaryReader | Uint8Array, length?: number): MsgCompleteDKG {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgCompleteDKG();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.sender = reader.string();
          break;
        case 2:
          message.id = reader.uint64();
          break;
        case 3:
          message.pubKeys.push(reader.string());
          break;
        case 4:
          message.consensusPubkey = reader.string();
          break;
        case 5:
          message.signature = reader.string();
          break;
        default:
          reader.skipType(tag & 7);
          break;
      }
    }
    return message;
  },
  fromPartial(object: Partial<MsgCompleteDKG>): MsgCompleteDKG {
    const message = createBaseMsgCompleteDKG();
    message.sender = object.sender ?? "";
    message.id = object.id !== undefined && object.id !== null ? BigInt(object.id.toString()) : BigInt(0);
    message.pubKeys = object.pubKeys?.map(e => e) || [];
    message.consensusPubkey = object.consensusPubkey ?? "";
    message.signature = object.signature ?? "";
    return message;
  },
  fromAmino(object: MsgCompleteDKGAmino): MsgCompleteDKG {
    const message = createBaseMsgCompleteDKG();
    if (object.sender !== undefined && object.sender !== null) {
      message.sender = object.sender;
    }
    if (object.id !== undefined && object.id !== null) {
      message.id = BigInt(object.id);
    }
    message.pubKeys = object.pub_keys?.map(e => e) || [];
    if (object.consensus_pubkey !== undefined && object.consensus_pubkey !== null) {
      message.consensusPubkey = object.consensus_pubkey;
    }
    if (object.signature !== undefined && object.signature !== null) {
      message.signature = object.signature;
    }
    return message;
  },
  toAmino(message: MsgCompleteDKG): MsgCompleteDKGAmino {
    const obj: any = {};
    obj.sender = message.sender === "" ? undefined : message.sender;
    obj.id = message.id !== BigInt(0) ? message.id.toString() : undefined;
    if (message.pubKeys) {
      obj.pub_keys = message.pubKeys.map(e => e);
    } else {
      obj.pub_keys = message.pubKeys;
    }
    obj.consensus_pubkey = message.consensusPubkey === "" ? undefined : message.consensusPubkey;
    obj.signature = message.signature === "" ? undefined : message.signature;
    return obj;
  },
  fromAminoMsg(object: MsgCompleteDKGAminoMsg): MsgCompleteDKG {
    return MsgCompleteDKG.fromAmino(object.value);
  },
  fromProtoMsg(message: MsgCompleteDKGProtoMsg): MsgCompleteDKG {
    return MsgCompleteDKG.decode(message.value);
  },
  toProto(message: MsgCompleteDKG): Uint8Array {
    return MsgCompleteDKG.encode(message).finish();
  },
  toProtoMsg(message: MsgCompleteDKG): MsgCompleteDKGProtoMsg {
    return {
      typeUrl: "/bitway.tss.MsgCompleteDKG",
      value: MsgCompleteDKG.encode(message).finish()
    };
  }
};
// ...
export const MsgSubmitSignatures = {
  typeUrl: "/bitway.tss.MsgSubmitSignatures",
  encode(message: MsgSubmitSignatures, writer: BinaryWriter = BinaryWriter.create()): BinaryWriter {
    if (message.sender !== "") {
      writer.uint32(10).string(message.sender);
    }
    if (message.id !== BigInt(0)) {
      writer.uint32(16).uint64(message.id);
    }
    for (const v of message.signatures) {
      writer.uint32(26).string(v!);
    }
    return writer;
  },
  decode(input: BinaryReader | Uint8Array, length?: number): MsgSubmitSignatures {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgSubmitSignatures();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.sender = reader.string();
          break;
        case 2:
          message.id = reader.uint64();
          break;
        case 3:
          message.signatures.push(reader.string());
          break;
        default:
          reader.skipType(tag & 7);
          break;
      }
    }
    return message;
  },
  fromPartial(object: Partial<MsgSubmitSignatures>): MsgSubmitSignatures {
    const message = createBaseMsgSubmitSignatures();
    message.sender = object.sender ?? "";
    message.id = object.id !== undefined && object.id !== null ? BigInt(object.id.toString()) : BigInt(0);
    message.signatures = object.signatures?.map(e => e) || [];
    return message;
  },
  fromAmino(object: MsgSubmitSignaturesAmino): MsgSubmitSignatures {
    const message = createBaseMsgSubmitSignatures();
    if (object.sender !== undefined && object.sender !== null) {
      message.sender = object.sender;
    }
    if (object.id !== undefined && object.id !== null) {
      message.id = BigInt(object.id);
    }
    message.signatures = object.signatures?.map(e => e) || [];
    return message;
  },
  toAmino(message: MsgSubmitSignatures): MsgSubmitSignaturesAmino {
    const obj: any = {};
    obj.sender = message.sender === "" ? undefined : message.sender;
    obj.id = message.id !== BigInt(0) ? message.id.toString() : undefined;
    if (message.signatures) {
      obj.signatures = message.signatures.map(e => e);
    } else {
      obj.signatures = message.signatures;
    }
    return obj;
  },
  fromAminoMsg(object: MsgSubmitSignaturesAminoMsg): MsgSubmitSignatures {
    return MsgSubmitSignatures.fromAmino(object.value);
  },
  fromProtoMsg(message: MsgSubmitSignaturesProtoMsg): MsgSubmitSignatures {
    return MsgSubmitSignatures.decode(message.value);
  },
  toProto(message: MsgSubmitSignatures): Uint8Array {
    return MsgSubmitSignatures.encode(message).finish();
  },
  toProtoMsg(message: MsgSubmitSignatures): MsgSubmitSignaturesProtoMsg {
    return {
      typeUrl: "/bitway.tss.MsgSubmitSignatures",
      value: MsgSubmitSignatures.encode(message).finish()
    };
  }
};
// ...
export const MsgCompleteRefreshing = {
  typeUrl: "/bitway.tss.MsgCompleteRefreshing",
  encode(message: MsgCompleteRefreshing, writer: BinaryWriter = BinaryWriter.create()): BinaryWriter {
    if (message.sender !== "") {
      writer.uint32(10).string(message.sender);
    }
    if (message.id !== BigInt(0)) {
      writer.uint32(16).uint64(message.id);
    }
    if (message.consensusPubkey !== "") {
      writer.uint32(26).string(message.consensusPubkey);
    }
    if (message.signature !== "") {
      writer.uint32(34).string(message.signature);
    }
    return writer;
  },
  decode(input: BinaryReader | Uint8Array, length?: number): MsgCompleteRefreshing {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgCompleteRefreshing();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.sender = reader.string();
          break;
        case 2:
          message.id = reader.uint64();
          break;
        case 3:
          message.consensusPubkey = reader.string();
          break;
        case 4:
          message.signature = reader.string();
          break;
        default:
          reader.skipType(tag & 7);
          break;
      }
    }
    return message;
  },
  fromPartial(object: Partial<MsgCompleteRefreshing>): MsgCompleteRefreshing {
    const message = createBaseMsgCompleteRefreshing();
    message.sender = object.sender ?? "";
    message.id = object.id !== undefined && object.id !== null ? BigInt(object.id.toString()) : BigInt(0);
    message.consensusPubkey = object.consensusPubkey ?? "";
    message.signature = object.signature ?? "";
    return message;
  },
  fromAmino(object: MsgCompleteRefreshingAmino): MsgCompleteRefreshing {
    const message = createBaseMsgCompleteRefreshing();
    if (object.sender !== undefined && object.sender !== null) {
      message.sender = object.sender;
    }
    if (object.id !== undefined && object.id !== null) {
      message.id = BigInt(object.id);
    }
    if (object.consensus_pubkey !== undefined && object.consensus_pubkey !== null) {
      message.consensusPubkey = object.consensus_pubkey;
    }
    if (object.signature !== undefined && object.signature !== null) {
      message.signature = object.signature;
    }
    return message;
  },
  toAmino(message: MsgCompleteRefreshing): MsgCompleteRefreshingAmino {
    const obj: any = {};
    obj.sender = message.sender === "" ? undefined : message.sender;
    obj.id = message.id !== BigInt(0) ? message.id.toString() : undefined;
    obj.consensus_pubkey = message.consensusPubkey === "" ? undefined : message.consensusPubkey;
    obj.signature = message.signature === "" ? undefined : message.signature;
    return obj;
  },
  fromAminoMsg(object: MsgCompleteRefreshingAminoMsg): MsgCompleteRefreshing {
    return MsgCompleteRefreshing.fromAmino(object.value);
  },
  fromProtoMsg(message: MsgCompleteRefreshingProtoMsg): MsgCompleteRefreshing {
    return MsgCompleteRefreshing.decode(message.value);
  },
  toProto(message: MsgCompleteRefreshing): Uint8Array {
    return MsgCompleteRefreshing.encode(message).finish();
  },
  toProtoMsg(message: MsgCompleteRefreshing): MsgCompleteRefreshingProtoMsg {
    return {
      typeUrl: "/bitway.tss.MsgCompleteRefreshing",
      value: MsgCompleteRefreshing.encode(message).finish()
    };
  }
};